+++ processor/src/test/resources/bad_input/com/example/default_props/InstanceComponent.java
@@ -1,0 +1,26 @@
+package com.example.default_props;
+
+import javax.annotation.Nonnull;
+import react4j.annotations.ReactComponent;
+import react4j.core.BaseContext;
+import react4j.core.BaseProps;
+import react4j.core.BaseState;
+import react4j.core.Component;
+import react4j.core.ReactNode;
+
+@ReactComponent
+class InstanceComponent
+  extends Component<BaseProps, BaseState, BaseContext>
+{
+  @Nonnull
+  BaseProps getInitialProps()
+  {
+    return new BaseProps();
+  }
+
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
+}