+++ processor/src/test/resources/expected/com/example/default_props/PropsSubclassDefaultPropsComponent_.java
@@ -1,0 +1,55 @@
+package com.example.default_props;
+
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+import jsinterop.base.JsPropertyMap;
+import react4j.core.BaseContext;
+import react4j.core.BaseProps;
+import react4j.core.BaseState;
+import react4j.core.ComponentConstructorFunction;
+import react4j.core.NativeAdapterComponent;
+import react4j.core.React;
+import react4j.core.ReactConfig;
+import react4j.core.ReactNode;
+
+@Generated("react4j.processor.ReactProcessor")
+class PropsSubclassDefaultPropsComponent_ extends PropsSubclassDefaultPropsComponent {
+  private static final ComponentConstructorFunction<BaseProps, BaseContext> TYPE = getConstructorFunction();
+
+  @Nonnull
+  static ReactNode _create() {
+    return React.createElement( TYPE );
+  }
+
+  @Nonnull
+  static ReactNode _create(@Nullable final BaseProps props) {
+    return React.createElement( TYPE, props );
+  }
+
+  @Nonnull
+  static ReactNode _create(@Nullable final BaseProps props, @Nullable final ReactNode child) {
+    return React.createElement( TYPE, props, child );
+  }
+
+  @Nonnull
+  private static ComponentConstructorFunction<BaseProps, BaseContext> getConstructorFunction() {
+    final ComponentConstructorFunction<BaseProps, BaseContext> componentConstructor = NativeReactComponent::new;
+    if ( ReactConfig.enableComponentNames() ) {
+      JsPropertyMap.of( componentConstructor ).set( "displayName", "PropsSubclassDefaultPropsComponent" );
+    }
+    JsPropertyMap.of( componentConstructor ).set( "defaultProps", PropsSubclassDefaultPropsComponent.getInitialProps() );
+    return componentConstructor;
+  }
+
+  static final class NativeReactComponent extends NativeAdapterComponent<BaseProps, BaseState, BaseContext, PropsSubclassDefaultPropsComponent> {
+    NativeReactComponent(@Nonnull final BaseProps props, @Nonnull final BaseContext context) {
+      super( props, context );
+    }
+
+    @Override
+    protected PropsSubclassDefaultPropsComponent createComponent() {
+      return new PropsSubclassDefaultPropsComponent_();
+    }
+  }
+}